Extract tree leaf component
